api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-pptx==1.0.2
openai==1.30.0
Pillow==10.0.1
aiofiles==23.2.1